requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
charset-normalizer==2.1.0
click==8.0.4
colorama==0.4.4
discord.py @ git+https://github.com/Rapptz/discord.py@b3a24846bddc894ae6d9bfb49ed8ee93b2d12fcb
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
python-dotenv==0.20.0
tomli==2.0.1
yarl==1.7.2